feat: Add HEPData probability models for pyhf regression tests
For stability in regression tests and to remove dependency on HEPData
uptime and delivery add the probability models for INSPIRE entry
1748602 published to HEPData to scikit-hep-testdata for regression
tests in pyhf.

Additionally add a file that contains the DOI for the probability
model archive itself, which exists for this particulare HEPData
entry. This provides clear DOI metadata for the future.

c.f.:
* https://inspirehep.net/literature/1748602
* https://doi.org/10.17182/hepdata.89408
* https://doi.org/10.17182/hepdata.89408.v3/r2